% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loader.R
\name{check_delphi_utils}
\alias{check_delphi_utils}
\title{Checking proper loading of \code{delphi-utils}}
\usage{
check_delphi_utils()
}
\value{
None
}
\description{
Checking proper loading of \code{delphi-utils}
}
\details{
This function double checks whether \code{delphi-utils} is available.
If it is available, returns the version installed.
}
