% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loader.R
\name{install_delphi_utils}
\alias{install_delphi_utils}
\title{Installing \code{delphi-utils} using \code{reticulate}}
\usage{
install_delphi_utils(method = "auto", conda = "auto", ...)
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically
finds a method that will work in the local environment. "virtualenv" is not
available on Windows.}

\item{conda}{The path to conda executable. By default, "auto" will allow
\code{reticulate} to automatically find an appropriate conda-binary}

\item{...}{Additional arguments passed to \code{reticulate::py_install}. Since
\code{delphi-utils} is only available on PyPi, pip is always set to TRUE.}
}
\description{
Installing \code{delphi-utils} using \code{reticulate}
}
\seealso{
\link{py_install}
}
